==> 01_luminal_cell_selection.Rmd <==
# (PART) Búsqueda de biomarcadores {-}

# Selección de células luminales de prostata {#luminal-cells-selection}

En esta sección se identificarán y seleccionarán células luminales de próstata, precursoras de la enfermedad, que posteriormente se usarán para identificar biomarcadores de la enfermedad. Para ello se utilizarán los datos de experimentos de scRNA-seq obtenidos de muestras de pacientes con la enfermedad [@base].

```{r Librerias, include=FALSE}
library(Seurat)
library(dplyr)
library(purrr)
```

## Importar matrices de cuentas

El primer paso será importar los datos, es decir las matrices de cuentas de cada paciente. Para facilitar este paso se define la función `import_gene_counts` que realizará los siguientes pasos:

1. Usará la función `data.table::fread`, que permite una lectura rápida de archivos de gran tamaño.
1. Asignará a las filas el nombre de las variables (genes) con la función `rownames`, los cuales se encontraban almacenados en la primera columna.
1. Finalmente con la función `dplyr::select` se eliminará la primera columna con los nombres.

En este caso se trabaja con dos conjuntos de datos provenientes de dos pacientes distintos, por lo que la función se aplicará para cada uno de ellos.

```{r Importar el dataset, warning=FALSE}
import_gene_counts <- function(file, data_table = FALSE) {
    count_matrix <- data.table::fread(file, data.table = data_table)
    rownames(count_matrix) <- count_matrix[[1]]
    count_matrix <- select(count_matrix, -(1))
    count_matrix
}

GSM4773522_PCa2_counts_matrix <- import_gene_counts(
    "./GSE157703_RAW/GSM4773522_PCa2_gene_counts_matrix.txt",
    data_table = FALSE
)

GSM4773522_PCa1_counts_matrix <- import_gene_counts(
    "./GSE157703_RAW/GSM4773521_PCa1_gene_counts_matrix.txt",
    data_table = FALSE
)
```

## Preprocesamiento y CQ

Para realizar el análisis se utilizará el paquete **Seurat**, el cual proveerá las herramientas necesarias para llevar a cabo la selección de biomarcadores, así como de células de interés [@R-Seurat].

### Creación del objeto Seurat

Antes de proceder al análisis de datos en experimentos de sc-RNAseq es recomendable realizar un control de calidad de las células. Con este proceso, es posible eliminar células de baja calidad que pueden generar ruido y malinterpretaciones de los datos, e.g. células rotas o muertas, multipletes, etc. La comunidad ha descrito diversas métricas para filtras estas células [@qc-sc], pero de forma intuitiva podemos considerar las siguientes:

1. **Número de genes únicos expresados en cada célula**:
    + Las células de baja calidad presentan muy pocos genes únicos.
    + Los multipletes presentan una aberrante alta cantidad de cuentas de genes.
1. **Número total de lecturas en una célula**:
    + Correlacionado con la cantidad de genes únicos.
1. **Porcentaje de lecturas mapeadas al genoma mitocondrial**:
    + Las células de baja calidad/muertas tienen una elevada contaminación mitocondrial.

**Seurat** permite realizar este preprocesamiento durante la creación del objeto principal. En este caso, se filtrarán aquellos genes que se hayan expresado en menos de 3 células (`min.cells = 3`) y las células que hayan expresado menos de 200 genes distintos (`min.features = 200`). Estos valores se han decidido en función de los conjuntos de datos utilizados [@base], para otros contextos otros valores pueden seleccionarse.

La función para crear los objetos con los que se trabajará (`CreateSeuratObject`), utilizará las matrices de cuentas creadas en el apartado anterior y opcionalmente un identificador (`project`) que arbitrariamente utilizaremos para indicar el paciente del que provienen.

```{r Creación de objetos Seurat y filtrado, warning=FALSE}
seurat_object_1 <- CreateSeuratObject(
    counts = GSM4773522_PCa1_counts_matrix,
    min.cells = 3,
    min.features = 200,
    project = "Patient1"
)

seurat_object_2 <- CreateSeuratObject(
    counts = GSM4773522_PCa2_counts_matrix,
    min.cells = 3,
    min.features = 200,
    project = "Patient2"
)
```

Aunque para la creación de los objetos se parte de 2 conjuntos de datos distintos (uno por paciente), podemos fusionarlos en 1 solo, de forma que sea más sencillo trabajar con ellos. Esto es posible porque cada célula tiene asignada un identificador único que la diferencia del resto, además de indicar del paciente del que proviene. Para fusionar ambos objetos simplemente se utilizará la función `merge`.

```{r Combinando objetos}
seurat_objects <- merge(seurat_object_1, seurat_object_2)
```

### Lecturas mapeadas al genoma mitocondrial

Tras realizar un primer filtrado durante la creación de los objetos se debe realizar un segundo en función del porcentaje de lecturas mapeadas al genoma mitocondrial. **Seurat** no ofrece por defecto una forma de realizar este filtrado durante la creación del objeto, por lo que se realizará manualmente.

En primer lugar, se creará una métrica con el porcentaje de lecturas mapeadas al genoma mitocondrial, para ello se utilizará la función `PercentageFeatureSet`. Esta función calcula para cada célula el porcentaje total de lecturas que pertenecen a un subgrupo concreto de genes. En el conjunto de datos, los genes pertenecientes al genoma mitocondrial están nombrados usando las iniciales "MT", lo que se utilizará para indicar el subgrupo de genes con el argumento `pattern`.

```{r Cálculo del porcentaje de reads de genoma mitocondrial}
seurat_objects[["percent.mt"]] <- PercentageFeatureSet(
    seurat_objects,
    pattern = "^MT-"
)
```

Generada la métrica, se realizará el filtrado utilizando la función `subset`. Con ella, se seleccionará el subconjunto de los datos que cumpla con las condiciones especificadas, que para este caso, será de un porcentaje inferior a $0.4$ [@base].

```{r Filtrado por porcentaje mitocondrial}
seurat_objects <- subset(
    seurat_objects,
    subset = percent.mt < 4
)
```

### Normalización

Hasta este punto se ha trabajado con los datos en crudo, sin embargo para realizar análisis aguas abajo es conveniente trabajar con datos normalizados. Realizar una normalización de los datos nos permite que la heterogeneidad de los mismos se deba principalmente a diferencias biológicas y no a otros sesgos, que en su mayoría son técnicos e.g. diferencias en distintas células en la eficiencia  en captura de cDNAs, ampliación por PCR etc. [@osca].

Para realizar esta normalización, se utilizará la función la función `Seurat::NormalizeData` con el argumento `normalization.method = "LogNormalize"`. Esta función aplicará un escalado global, es decir: en cada célula dividirá el número de cuentas de cada gen entre el número total de cuentas. Finalmente aplicará una trasformación logarítmica a los resultados.

```{r Normalización, cache=TRUE, results='hide', message=FALSE}
seurat_objects <- NormalizeData(
    seurat_objects,
    normalization.method = "LogNormalize"
)
```

## Búsqueda de genes de alta variabilidad

Una vez realizado el preprocesamiento de los datos, se buscarán los genes que presentan una mayor variabilidad, es decir, que presenten niveles de expresión muy altos en algunas células y muy bajos en otras. La búsqueda de estos genes, y su uso en análisis aguas abajo, facilita la identificación de patrones de interés [@high-variability-genes], por este motivo los usaremos en los siguientes pasos.

Para seleccionar estos genes se usará la función `Seurat::FindVariableFeatures`, la cual requerirá, además del objeto `seurat`, un número de variables a seleccionar (`nfeatures`) que para este caso serán $2000$.

```{r Búsqueda de variables, cache=TRUE, warning=FALSE, message=FALSE}
seurat_objects <- FindVariableFeatures(
    seurat_objects,
    selection.method = "vst",
    nfeatures = 2000
)
```

Las variables seleccionadas por la función pueden visualizarse utilizando `Seurat::VariableFeaturePlot` en función de su media y varianza. Para mayor claridad, los 10 primeros genes se mostrarán con su nombre en la figura \@ref(fig:variable-genes).

```{r variable-genes, warning=FALSE, fig.cap="Genes de alta variabilidad", message=FALSE}
high_variable_genes <- head(VariableFeatures(seurat_objects), 10)
high_variable_genes_plot <- VariableFeaturePlot(seurat_objects)
LabelPoints(
    plot = high_variable_genes_plot,
    points = high_variable_genes,
    repel = TRUE
)
```

## Escalado de los datos

En este tipo de experimentos, donde se trabaja con un gran número de variables, es habitual aplicar técnicas de reducción de la dimensionalidad. Sin embargo, previamente a este tipo de técnicas es conveniente aplicar un escalado y centrado de los datos. Esto se debe a que este tipo de técnicas son sensibles a conjuntos de datos donde las variables presentan escalas muy dispares. Realizando un escalado, genes que habitualmente se encuentren mucho más expresados de forma basal, no tendrán más peso en el análisis.

**Seurat** provee la función `ScaleData`, que realizará este escalado además de un centrado de los datos. Por defecto la función hará que la expresión media para cada gen en todas las células sea $0$ y su varianza igual a $1$. Finalmente se indicará que los cálculos solo se realicen sobre los genes de alta variabilidad (`features`), dado que son los que se utilizarán posteriormente para el análisis.

```{r Escalado, warning=FALSE, message=FALSE}
seurat_objects <- ScaleData(
    seurat_objects,
    features = VariableFeatures(seurat_objects)
)
```

## Reducción de la dimensionalidad {#dim-reduction}

Como se ha expresado previamente, el conjunto de datos tiene un alto número de variables y por tanto es conveniente aplicar técnicas de reducción de la dimensionalidad para trabajar con él. Por este motivo se realizará un análisis de componentes principales (PCA) sobre los datos ya escalados.

Para aplicar esta técnica, **Seurat** provee la función `RunPCA` donde, de forma similar a la sección anterior, se especificarán los genes de mayor variabilidad (`features`), dado que serán estos con los que se desea trabajar.

```{r PCA, cache=TRUE, warning=FALSE, message=FALSE}
seurat_objects <- RunPCA(
    seurat_objects,
    features = VariableFeatures(seurat_objects)
)
```

## Integración de datos

### Objetivos de la integración

La integración de datos en scRNA-seq permite harmonizar conjuntos de datos dispares en una sola referencia [@integration]. Expresado de otra forma y aplicado a este caso, la integración permite encontrar correspondencias entre las células de ambos conjuntos de datos. De esta forma, es posible proyectar los resultados de un experimento en el espacio del otro, solventando diferencias que puedan deberse a utilizar datos de provenientes de pacientes distintos. Así, células del mismo tipo deberían agruparse juntas independientemente del individuo del que provengan.

Indirectamente, la integración permite mejorar el poder estadístico en análisis posteriores, dado que permite utilizar varios conjuntos de datos simultáneamente. Además, aunque algunos de estos casos no se aplique para este trabajo, también habilita:

+ Identificar subpoblaciones de células presentes en conjuntos de datos distintos.
+ Comparar marcadores específicos de tipos celulares en diferentes condiciones (e.g. control vs tratamiento).

### Integración

Existen diversos métodos que se pueden utilizar para realizar la integración de datos, pero en este caso se usará el *análisis de correlación canónica* (CCA) [@integration].

Cuando se aplica este método, **Seurat** trabaja en un espacio de baja dimensionalidad, es decir usando las dimensiones generadas previamente en el PCA. A estas nuevas dimensiones se les aplica unas correcciones para generar unas nuevas que contemplan las fuentes de variación entre ambas muestras, de esta forma células del mismo tipo se agruparán posteriormente independientemente del paciente del que provengan.

Para realizar esta integración se utilizará la función `IntegrateLayers`. En ella simplemente se indicará el nombre de la nueva reducción (`new.reduction`), la reducción a corregir (`orig.reduction`) y el método de integración a utilizar (`method = CCAIntegration`).

```{r Integración, warning=FALSE, cache=TRUE, message=FALSE}
seurat_objects <- IntegrateLayers(
    seurat_objects,
    method = CCAIntegration,
    orig.reduction = "pca",
    new.reduction = "integrated.cca"
)
```

Realizada la integración, los conjuntos de datos del objeto pueden finalmente fusionarse para proseguir los análisis aguas abajo. Seurat realiza este paso internamente con la función `JoinLayers`.

```{r Unir layers}
seurat_objects <- JoinLayers(seurat_objects)
```

## Dimensionalidad del conjunto de datos

Como se ha expresado en la sección \@ref(dim-reduction), los datos de experimentos de scRNA-seq presentan un gran número de variables y ruido. Por este motivo, trabajar con las variables de mayor variabilidad y aplicar técnicas para reducir el número de dimensiones, como el PCA, es recomendable.

Las componentes producidas por esta técnica representan un conjunto de variables iniciales que están **correlacionadas entre sí**. Además, las variables de componentes distintas **no están correlacionadas**, por este motivo cada componente describe una cantidad de la variabilidad total del conjunto de datos. De esta forma, se plantea la cuestión de cuántas componentes se deben utilizar en análisis posteriores, o la "dimensionalidad" del conjunto de datos. Utilizar un gran número de componentes encapsula una mayor variabilidad de la total, sin embargo dificulta la tarea por el elevado número de variables. Por otra parte, trabajar con un menor número de dimensiones será más sencillo pero conservará una menor variabilidad, perdiendo información de esta forma.

Para solventar este problema existen distintos métodos que ayudan a seleccionar la dimensionalidad de un conjunto de datos. Un método heurístico comúnmente utilizado es 'el método del codo', el cual implica representar gráficamente la variación explicada en función de su componente. Seurat provee la función `ElbowPlot` para aplicar este método.

```{r elbow-plot, fig.cap="Método del codo. En ella se muestra varición explicada por una de las componentes principales"}
ElbowPlot(seurat_objects, ndims = 50)
```

Otra forma de visualizar esta variabilidad es a través de mapas de calor de las componentes principales. Seurat también permite automatizar este proceso a través de la función `DimHeatmap`.

```{r heatmaps, fig.cap="Mapas de calor de las 21 primeras componentes"}
DimHeatmap(
    seurat_objects,
    dims = 1:21,
    cells = 500,
    balanced = TRUE
)
```

Tras observar la figura \@ref(fig:elbow-plot) y \@ref(fig:heatmaps), se puede observar que la variación empieza a disminuir a partir de la componente número 10. Sin embargo, se debe tener en cuenta que escoger menos componentes de lo ideal puede resultar en perdidas de información y propagación a los resultados finales. Por este motivo y de forma arbitraria, se seleccionará hasta la componente 20 donde el descenso de la variación explicada es más pronunciado, o en otras palabras, se está capturando la mayor parte de la variabilidad de los datos.

## Agrupamiento de células {#clustering}

Seleccionadas las dimensiones, el siguiente paso consistirá en agrupar las células del experimento. En primer lugar, se construirá un grafo con los K vecinos más cercanos (KNN) utilizando las dimensiones seleccionadas previamente. Para ello, se usará la función `Seurat::FindNeighbors`.

```{r K vecinos más cercanos, cache=TRUE, message=FALSE}
seurat_objects <- FindNeighbors(
    seurat_objects,
    dim = 1:20,
    reduction = "integrated.cca"
)
```

Generado el grafo, **Seurat** realiza una optimización del paso anterior para finalmente agrupar las distintas células en comunidades. Para ello utiliza la función `FindClusters`, la cual además provee el parámetro `resolution` para controlar la granularidad del agrupamiento. Para el rango de células en el que estamos trabajando utilizar valores entre $0.4$ y $1.2$ es habitual. De forma arbitraria, en este caso se utilizará una resolución de $0.8$.

```{r Búsqueda de cluster, cache=TRUE, message=FALSE, results='hide'}
seurat_objects <- FindClusters(
    seurat_objects,
    resolution = 0.8
)
```

## Reducción de la dimensionalidad no lineal (UMAP)

### Utilidad del UMAP

Las técnicas de reducción de la dimensionalidad no lineales pueden mostrar relaciones subyacentes en los datos que no hayan sido capturadas por técnicas lineales. En este caso, la técnica *uniform manifold approximation and projection* (UMAP) puede revelar alguna de estas relaciones no identificadas previamente por el PCA.

Aunque esta técnica puede revelar relaciones ocultas, también se han de considerar sus limitaciones: este método preserva las distancias locales del conjunto de datos, a costa de ignorar las de mayor distancia. De esta forma, aunque la técnica puede revelar relaciones ocultas, no es recomendable utilizarla para hacer conclusiones, dado que muchas otras no se habrán considerado.

### UMAP

Es posible realizar un UMAP con la función `Seurat::RunUMAP` indicando las dimensiones a utilizar (`dims`) y la reducción de la dimensionalidad que queremos que utilice (`reduction`), en este caso en la que se han integrado los datos (`integrated.cca`)

```{r UMAP, cache=TRUE, message=FALSE, warning=FALSE}
seurat_objects <- RunUMAP(
    seurat_objects,
    dims = 1:20,
    reduction = "integrated.cca"
)
```

### Visualización del UMAP

Para visualizar las dimensiones generadas es posible utilizar la función `Seurat::DimPlot`. En ella simplemente se habrá de indicar la reducción que se quiere mostrar (`reduction`), en este caso el UMAP.

```{r umap-view, fig.cap="Clusters en las dimensiones generadas por el UMAP"}
DimPlot(
    seurat_objects,
    reduction = "umap",
    label = TRUE
)
```

## Búsqueda de genes expresados diferencialmente (DEG) {#degs-search}

Tras identificar los grupos de células es posible buscar los genes diferencialmente expresados en cada uno de ellos. Para esta tarea **Seurat** provee la función `FindAllMarkers` que busca los genes expresados en mayor o menor medida con respecto al resto de células que no pertenecen al cluster.

```{r DEG en clusters, cache=TRUE, message=FALSE}
all_markers <- FindAllMarkers(
    seurat_objects
)
```

La función devuelve una tabla con las siguientes columnas:

+ **p_val**\. Valor p para un gen al realizar un test de Suma de rangos de Wilcoxon.
+ **avg_log2FC**\. El log fold-change entre los dos grupos. El primer grupo se corresponderá con el cluster, el segundo con el resto de células que no pertenecen a él.
+ **pct.1**\. El porcentaje de células en el cluster que expresa el gen.
+ **pct.2**\. El porcentaje de células que expresan el gen y que no pertenecen al cluster.
+ **p_val_adj**\. El valor p ajustado por la corrección de Bonferroni usando todos los genes del conjunto de datos.

## Selección de células

### Filtrado de resultados significativos

Obtenidos los DEGs y los clusters a los que pertenecen se pueden analizar los resultados, pero antes se habrán de seleccionar aquellos que sean significativos. Dado que se están realizando un gran número de pruebas (una por cada gen), se hará uso del p-valor ajustado por la corrección de Bonferroni, calculado en la sección anterior.

Para realizar el filtrado de los resultados que cumplan la condición $\text{p-valor} \le 0.05$, se utilizará la función `dplyr::filter`.

```{r Filtrado por valores p}
significative_markers <- all_markers %>%
    filter(p_val_adj <= 0.05)
```

### Identificación de clusters

Seleccionados los DEGs significativos, se pueden identificar los tipos celulares de cada cluster. En concreto, como en este caso se está trabajando con muestras de cáncer de próstata, la búsqueda se centrará en encontrar las células propias de este tejido, filtrando posibles contaminaciones.

En otros estudios se han reportado marcadores genéticos de células luminales de próstata [@luminal-markers], las contribuidoras principales al desarrollo del tumor y de la enfermedad. Los principales marcadores en este tipo de células son "Keratin 8" (*KRT8*) y "Keratin 18" (*KRT18*) por lo que se puede explorar si alguno de los cluster los expresa.

```{r Identificación de clusters, eval=FALSE}
significative_markers %>%
    filter(gene == "KRT18" | gene == "KRT8") %>%
    arrange(desc(avg_log2FC), desc(avg_log2FC))
```

```{r, echo=FALSE}
knitr::kable(
    significative_markers %>%
        filter(gene == "KRT18" | gene == "KRT8") %>%
        arrange(desc(avg_log2FC), desc(avg_log2FC)) %>%
        mutate(across(c(p_val, p_val_adj), ~ tibble::num(.x, digits = 3, notation = "sci")))
)
```

Al ordenar los resultados por su `avg_log2FC` y `pct.1`, con la función `dplyr::arrange`, se observa que tanto *KRT8* y *KR18* se expresan en los mismos clusters. En concreto, en los clusters 8 y 14, los genes se encuentran en un alto porcentaje de las células y además con niveles de expresión relativamente altos. Para el caso del cluster 7, aunque presente niveles de expresión ligeramente superiores a los del cluster 8, no será seleccionado al presentar un porcentaje de expresión detectable muy bajo.

Estos resultados pueden observarse de forma gráfica utilizando la función `Seurat::FeaturePlot` especificando los genes que se que quieren mostrar.

```{r feature-krt, fig.cap="Expresión de KRT8 y KRT18 en los clusters 8 y 14"}
FeaturePlot(
    seurat_objects,
    feature = c("KRT8", "KRT18")
)
```

De esta forma, se deduce que los clusters seleccionados se corresponden con células luminales de próstata y por tanto serán objeto de estudio. Por este motivo, se seleccionarán las células pertenecientes a estos clusters para analizarlas en pasos posteriores.

```{r Selección de clusters}
luminal_cells <- subset(
    seurat_objects,
    idents = c(8, 14)
)
```

```{r Serialización del objeto, include=FALSE}
saveRDS(
    object = luminal_cells,
    file = "./rds_objects/01_luminal_object.rds"
)

saveRDS(
    object = significative_markers,
    file = "./rds_objects/01_significative_markers.rds"
)
```
