==> 04_Common_markers_selection.Rmd <==
# Filtrado de la matriz de conteo {#count-matrix-filtering}

En esta parte del trabajo se realizará un filtrado de la matriz generada en la sección anterior (sección \@ref(count-matrix-generation)). Para este paso, se utilizarán los biomarcadores seleccionados de experimentos de scRNA-seq (sección \@ref(cancer-cells-selection)), reduciendo así el tamaño y ruido presente en los datos lo que facilitará los análisis posteriores.

```{r Importar librerías, include=FALSE, warning=FALSE}
library(biomaRt)
library(dplyr)
library(tidyr)
library(stringr)
```

```{r Importar datos, include=FALSE}
markers_cluster_8 <- readRDS("./rds_objects/02_markes_cluster_8.rds")
tcga_data <- readRDS("./rds_objects/03_tcga_data.rds")
```

## Equivalencias entre identificadores

Los formas para identificar los biomarcadores en los experimentos de scRNA-seq y los provenientes de TCGA son distintas. Los primeros utilizan el acrónimo del gen, mientras que los segundos utilizan identificadores de [Ensembl](https://www.ensembl.org/index.html). 

Para realizar el filtrado se deberán encontrar equivalencias entre los acrónimos e identificadores de Ensembl. Esta tarea es posible utilizando el paquete `biomaRt`[@R-biomaRt], el cual permite acceder a diversos portales de información entre los que se encuentra Ensembl. En primer lugar, se usará la función `biomaRt::useEnsembl` para seleccionar la base de datos a la que acceder, en este caso `genes`, y el conjunto de datos a utilizar, en este caso `hsapiens_gene_ensembl`, es decir genes de humanos.

```{r Búsqueda de equivalencias para los identificadores, cache=TRUE}
ensembl <- useEnsembl(
    biomart = "genes",
    dataset = "hsapiens_gene_ensembl"
)
```

La función preparará la configuración necesaria para conectar con Ensembl y realizar consultas a su base de datos, para esto último se usará la función `biomaRt::getBM`. `getBM` permite especificar una serie de atributos para la consulta, en este caso, se requerirá el nombre del gen (`hgnc_symbol`), su identificador de Ensembl (`ensembl_gene_id`) y posibles nombres alternativos para dicho gen (`external_synonym`).

En la configuración de la función también se ha especificado el parámetro `mart`, que simplemente requiere el objeto con la configuración antes generada. Además de estos parámetros, la función también permite especificar filtros para recibir resultados más concretos. Habitualmente serían de utilidad, pero dado que para este caso se requerirá una lógica más compleja, por ello, en su lugar, se aplicarán filtrados propios en pasos posteriores. 

```{r Ejecutando la consulta, cache=TRUE}
query_results <- getBM(
    attributes = c(
        "hgnc_symbol",
        "ensembl_gene_id",
        "external_synonym"
    ),
    mart = ensembl
)
```
```{r, echo=FALSE}
knitr::kable(
    head(query_results)
)
```

Obtenidos los resultados de la consulta, antes de comenzar a trabajar con ellos, se compactará la columna de los nombres y sinónimos en una sola. De esta forma, será posible acceder a cualquier nombre y/o equivalente de forma más sencilla.

```{r Tabla de equivalencias}
id_equivalences <- query_results %>%
    pivot_longer(
        cols = !ensembl_gene_id,
        names_to = "type",
        values_to = "acronym"
    ) %>% distinct()
```
```{r head-id-equivalences, echo=FALSE}
kableExtra::kbl(
    head(id_equivalences)
) %>% kableExtra::kable_styling()
```

Generada la tabla de equivalencias, es conveniente comprobar que los biomarcadores seleccionados se encuentran en ella. De esta de forma, se asegura que sea posible asignarles un identificador. Realizando la intersección (`dplyr::intersect`) y la diferencia (`dplyir::setdiff`) entre los biomarcadores seleccionados y genes en la tabla de equivalencias, es posible identificar los genes que se han encontrado en la base de datos y los faltantes.

```{r Genes encontrados y no encontrados}
missing_genes <- setdiff(markers_cluster_8$gene, id_equivalences$acronym)
finded_genes <- intersect(markers_cluster_8$gene, id_equivalences$acronym)
```
```{r finded-missing-genes-datatable, echo=FALSE}
search_genes <- dplyr::bind_rows(
    tibble::tibble(gene = finded_genes, missing = FALSE),
    tibble::tibble(gene = missing_genes, missing = TRUE)
)

DT::datatable(
    search_genes,
    filter = "top",
    options = list(
        dom = "tip"
    )
)
```

Una exploración de los genes faltantes en la base de datos revela una posible explicación para esta ausencia. Tomando como ejemplo el caso de **RP11-356O9.1**, una búsqueda en ensamblados anteriores ([GRCh37.p13](https://grch37.ensembl.org/Homo_sapiens/Gene/Summary?db=core;g=ENSG00000258414;r=14:38033252-38048330;t=ENST00000554829)) permiten identificarlo por su nombre, sin embargo en casos posteriores ([GRCh38.p14](https://www.ensembl.org/Homo_sapiens/Gene/Idhistory?g=ENSG00000258414)) no existe nombre alguno para dicho gen.

Como sucede con este gen, puede darse el caso de correcciones de errores, es decir genes que inicialmente fueron identificados como distintos se combinan en una sola entrada cuando se descubre que puede tratarse del mismo. Alternativamente, el símbolo otorgado a estos genes se corresponde al establecido por [HUGO Gene Nomenclature Commitee](https://www.genenames.org/), por ello, en caso que un gen no haya recibido todavía un símbolo, este no estará disponible en la base de datos.

Por estos casos, para incluir correcciones más recientes se usará el ensamblado más reciente (GRCh38.14). Además, en filtrados posteriores, los genes que presenten cierta arbitrariedad no serán considerados a fin de evitar posibles errores.

## Filtrado por los genes identificados

Seleccionados los genes con identificadores, se obtendrán aquellos que pertenezcan a los datos de interés. Además, para facilitar posteriores accesos, se añadirán los identificadores de Ensembl al conjunto de datos en una nueva variable (`ensembl_gene_id`).

```{r Genes filtrados}
finded_markers <- markers_cluster_8 %>%
    filter(gene %in% finded_genes) %>%
    left_join(id_equivalences, by = join_by(gene == acronym))
```

Es posible observar que cuando esta nueva columna es creada, se obtiene un mayor número de entradas del esperado.

```{r Entradas duplicadas, eval=FALSE}
finded_markers %>%
    group_by(ensembl_gene_id) %>%
    filter(n() > 1) %>%
    arrange(ensembl_gene_id)
```
```{r duplicated-entries-kable, echo=FALSE}
test_markers <- finded_markers %>%
    group_by(ensembl_gene_id) %>%
    filter(n() > 1) %>%
    arrange(ensembl_gene_id)


kableExtra::kbl(
    finded_markers %>%
        group_by(ensembl_gene_id) %>%
        filter(n() > 1) %>%
        arrange(ensembl_gene_id) %>%
        mutate(across(c(p_val, p_val_adj), ~ tibble::num(.x, digits = 3, notation = "sci")))
) %>% kableExtra::kable_paper()
```

Comprobando las distintas entradas es posible observar como algunos genes presentan el mismo identificador de Ensembl. Estas duplicaciones han sido generadas durante la creación de la columna `ensembl_gene_id`, debido a uno de los siguientes casos:

+ El sinónimo (`external_sinonym`) y nombre del gen (`hgnc_symbol`) son iguales, introduciendo información duplicada (e.g. **PEX11G**).
+ El sinónimo y nombre del gen son distintos, introduciendo dos entradas (e.g. **APPL2** y **APLP2**).

Para solventar esta redundancia en la información, se realizará un filtrado siguiendo los siguientes criterios:

+ **Si un `ensembl_gene_id` es único**\. La entrada se conservará (`n() == 1)`).
+ **Si un `ensembl_gene_id` no es único**\. Entre las entradas que usen dicho identificador, se conservará el nombre del gen y descartarán posibles sinónimos (`type == "hgnc_symbol"`). En caso de que todas las entradas sean sinónimos se seleccionará, de forma arbitraria, el primer sinónimo que aparezca en los datos.

```{r Filtrado de duplicados}
filtered_markers <- finded_markers %>%
    group_by(ensembl_gene_id) %>%
    filter(
        case_when(
            n() == 1 ~ TRUE,
            n() > 1 & (type == "hgnc_symbol") ~ TRUE,
            n() > 1 & (gene == gene[[1]]) & all(cumall(type == "external_synonym")) ~ TRUE,
            .default = FALSE
        )
    )
```

Una exploración más detallada de los datos muestra el caso opuesto, genes con el mismo nombre presentando 2 identificadores de ensemble distintos, e.g. **PRR4** con los identificadores: `ENSG00000111215`, `ENSG00000263247` y `ENSG00000282269`. Esta duplicidad es conveniente, puesto que es posible asegurar a que identificador se hace referencia en los experimentos. Además, no afecta negativamente, puesto que aunque posean identificadores distintos los valores de expresión para estos genes serán idénticos.

Filtrados los biomarcadores, se podrán seleccionar los genes del conjunto de datos extraído de TCGA. Para ello, en primer lugar, se renombrarán las columnas eliminando la versión de los identificadores de ellas y los genes conteniendo "PAR_Y" en su identificador, puesto que no están expresados y pueden generar conflictos de identificadores duplicados. En segundo lugar, se seleccionarán las columnas (genes) que coincidan con los identificadores seleccionados.

```{r TCGA subsetting}
roc_data <- tcga_data %>%
    select(
        -contains("PAR")
    ) %>%
    rename_with(
        \(x) str_replace(x, "\\..*$", "")
    ) %>%
    select(
        contains(filtered_markers$ensembl_gene_id) |
        c("sample", "gleason_score", "disease", "malignancy", "prognostic")
    )
```

De esta forma, el conjunto de datos resultante contiene 2651 variables (exceptuando variables respuesta e identificadores de muestra). Como se ha explicado previamente, es posible que existan genes con identificadores distintos, sin embargo, se puede realizar un conteo del número de genes únicos que en realidad se han seleccionado.

```{r test}
genes_in_tcga <- filtered_markers$ensembl_gene_id %in%
    colnames(roc_data)
markers_in_tcga <- filtered_markers[which(genes_in_tcga), ]
length(unique(markers_in_tcga$gene))
```

En pasos anteriores, cuando se seleccionaron los biomarcadores comunes entre los obtenidos en las muestras y los disponibles en Ensembl, se encontraron 2585; y de estos, 2584 se han encontrado en las muestras del TCGA. Con estos genes, será con los que se trabaje a la hora de realizar los análisis posteriores.

```{r Serialización, include=FALSE}
saveRDS(
    object = markers_in_tcga,
    file = "./rds_objects/04_common_markers.rds"
)

saveRDS(
    object = roc_data,
    file = "./rds_objects/04_filtered_tcga_data.rds"
)
```
